Rename file.
