+++ lib/kavun.js
@@ -8,7 +8,7 @@ startTimer(processId);
 
 const runner = require('./runner');
 const { unitCollector } = require('./index');
-const { loadTestFiles } = require('./file-loader');
+const { loadTestFiles } = require('./test-file-loader');
 
 const path = process.argv[2];
 loadTestFiles(require, `${process.cwd()}/${path}`);
