+++ lib/test-file-loader.js
